refactor(go): split entrypoints and move bulky test bundle off main files
